+++ packages/Angular/Generic/agent-requests/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @memberjunction/ng-agent-requests
 
+## 6.0.0
+
+### Patch Changes
+
+- Updated dependencies [a2670a9]
+  - @memberjunction/core@6.0.0
+  - @memberjunction/ai-core-plus@6.0.0
+  - @memberjunction/ng-notifications@6.0.0
+  - @memberjunction/ng-shared-generic@6.0.0
+  - @memberjunction/core-entities@6.0.0
+  - @memberjunction/ng-forms@6.0.0
+  - @memberjunction/global@6.0.0
+
 ## 5.51.0
 
 ### Patch Changes