+++ packages/Angular/Generic/word-cloud/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @memberjunction/ng-word-cloud
 
+## 6.0.0
+
 ## 5.51.0
 
 ## 5.50.0